Addressed code that assumed the long integer type was 4 bytes when converting to/from the float type.  For 64-bit systems, long is 8 bytes.  Changed code to use integer type, which is 4 bytes in both 32 and 64-bit systems.

+++ branches/ssg/sdl/headers/externs.h
@@ -298,7 +298,7 @@ extern S32                   maxct;
 extern char                  maxfn;
 extern long                  maxit;
 extern int                   maxlinelength;
-extern long                  MaxLTSize;
+extern int                  MaxLTSize;
 extern unsigned              Max_Args;
 extern unsigned              Max_Ops;
 extern long                  maxptr;
